requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
slowapi==0.1.9
asyncpg==0.29.0
